Integration test cleanup, make registering tests much easier Update junit to latest version
