+++ pom.xml
@@ -52,7 +52,7 @@
     </distributionManagement>
 
     <properties>
-        <junit.jupiter.version>5.0.1</junit.jupiter.version>
+        <junit.jupiter.version>5.4.2</junit.jupiter.version>
         <junit.platform.version>1.0.1</junit.platform.version>
         <guava.version>23.0</guava.version>
         <mockito.version>1.9.0</mockito.version>
