+++ src/test/java/com/whiteclarkegroup/liquibaselinter/integration/LinterIntegrationTest.java
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.TestFactory;
 import org.junit.jupiter.api.function.ThrowingConsumer;
 
 import java.io.Writer;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 
@@ -20,10 +21,13 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 abstract class LinterIntegrationTest {
 
-    abstract List<IntegrationTestConfig> getTests();
+    private final List<IntegrationTestConfig> tests = new ArrayList<>();
+
+    abstract void registerTests();
 
     @TestFactory
     Stream<DynamicTest> dynamicTests() {
+        registerTests();
         ThrowingConsumer<IntegrationTestConfig> testExecutor = running -> {
             final Liquibase liquibase = LiquibaseIntegrationTestResolver.buildLiquibase(running.getChangeLogFile(), running.getConfigFile());
             final Writer nullWriter = CharStreams.nullWriter();
@@ -36,7 +40,7 @@ abstract class LinterIntegrationTest {
                 liquibase.update(contexts, nullWriter);
             }
         };
-        return DynamicTest.stream(getTests().iterator(), IntegrationTestConfig::getDisplayName, testExecutor);
+        return DynamicTest.stream(tests.iterator(), IntegrationTestConfig::getDisplayName, testExecutor);
     }
 
     @AfterEach
@@ -44,4 +48,41 @@ abstract class LinterIntegrationTest {
         ChangeLogParserFactory.reset();
     }
 
+    protected void shouldFail(String displayName, String changeLogFile, String configFile, String message) {
+        tests.add(new IntegrationTestConfig(displayName, changeLogFile, configFile, message));
+    }
+
+    protected void shouldPass(String displayName, String changeLogFile, String configFile) {
+        tests.add(new IntegrationTestConfig(displayName, changeLogFile, configFile, null));
+    }
+
+    private static class IntegrationTestConfig {
+        private final String displayName;
+        private final String changeLogFile;
+        private final String configFile;
+        private final String message;
+
+        private IntegrationTestConfig(String displayName, String changeLogFile, String configFile, String message) {
+            this.displayName = displayName;
+            this.changeLogFile = changeLogFile;
+            this.configFile = configFile;
+            this.message = message;
+        }
+
+        private String getDisplayName() {
+            return displayName;
+        }
+
+        private String getChangeLogFile() {
+            return changeLogFile;
+        }
+
+        private String getConfigFile() {
+            return configFile;
+        }
+
+        private String getMessage() {
+            return message;
+        }
+    }
 }
